Add home and footer
