+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "pretty": "npx prettier --write ."
   },
   "dependencies": {
     "@emotion/react": "^11.8.2",
