+++ styles/globals.css
@@ -16,6 +16,6 @@ a {
   box-sizing: border-box;
 }
 
-.MuiContainer-maxWidthXs	{
+.MuiContainer-maxWidthXs {
   max-width: 8rem;
 }
